[update] 解决扫描过期时间的bug

+++ main.go
@@ -55,13 +55,16 @@ func main() {
 				}
 				for _, user := range users {
 					// 如果过期时间在现在之前，则删除
-					if user.ExpiredTime.Before(time.Now()) {
-						// 删除user
-						_, err := service.DeleteUSer(user.ID)
-						if err != nil {
-							zap.S().Errorln("删除user失败:", err)
-						} else {
-							zap.S().Infoln("delete user:", user.Username)
+					nilTime := time.Time{}
+					if user.ExpiredTime != nilTime {
+						if user.ExpiredTime.Before(time.Now()) {
+							// 删除user
+							_, err := service.DeleteUSer(user.ID)
+							if err != nil {
+								zap.S().Errorln("删除user失败:", err)
+							} else {
+								zap.S().Infoln("delete user:", user.Username)
+							}
 						}
 					}
 				}
